PUMP: Power Under Multiplicity Project

Estimates power, minimum detectable effect size (MDES) and sample size requirements. The context is multilevel randomized experiments with multiple outcomes. The estimation takes into account the use of multiple testing procedures. Development of this package was supported by a grant from the Institute of Education Sciences (R305D170030). For a full package description, including a detailed technical appendix, see doi:10.18637/jss.v108.i06 .

See also

Author

Maintainer: Luke Miratrix luke_miratrix@gse.harvard.edu (ORCID)

Authors:

Kristen Hunter kristen.hunter@unsw.edu.au (ORCID)

Zarni Htet

Kristin Porter kristin.porter@keporterconsulting.com

Other contributors:

MDRC [copyright holder]

Institute of Education Sciences [funder]